webserver/requirements.txt
flask
Flask-WTF
Flask-Login

# native driver; see https://neo4j.com/docs/api/python-driver/current/api.html
neo4j

# https://github.com/allofphysicsgraph/proofofconcept/issues/157
secure==0.3.0

# Sign in with Google
google-api-python-client
# https://google-auth.readthedocs.io/en/latest/reference/google.auth.transport.requests.html
google-auth-httplib2
google-auth-oauthlib

# only necessary for localhost
# from https://blog.miguelgrinberg.com/post/running-your-flask-application-over-https
#pyOpenSSL==24.2.1
pyOpenSSL

# image size
opencv-python

# code formatting and consistency checking
#mypy==0.910
mypy
#black==21.7b0
black
#flake8==3.8.4
flake8
#pylint==2.5.3
pylint

# step validation
#sympy==1.12
sympy
antlr4-python3-runtime==4.11

# only necessary for localhost
# from https://blog.miguelgrinberg.com/post/running-your-flask-application-over-https
#pyOpenSSL==20.0.1
pyOpenSSL

#EOF
